# users/server/models/get_friendship_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// GetFriendshipResponse get friendship response
//
// swagger:model GetFriendshipResponse
type GetFriendshipResponse struct {

	// followed id
	// Example: 2
	FollowedID int64 `json:"followed_id,omitempty"`

	// follower id
	// Example: 1
	FollowerID int64 `json:"follower_id,omitempty"`

	// id
	// Example: 5
	ID int64 `json:"id,omitempty"`

	// status
	// Example: accepted
	Status string `json:"status,omitempty"`
}

// Validate validates this get friendship response
func (m *GetFriendshipResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this get friendship response based on context it is used
func (m *GetFriendshipResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *GetFriendshipResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GetFriendshipResponse) UnmarshalBinary(b []byte) error {
	var res GetFriendshipResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}