# users/server/models/login_user_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// LoginUserResponse login user response
//
// swagger:model LoginUserResponse
type LoginUserResponse struct {

	// token
	Token string `json:"token,omitempty"`
}

// Validate validates this login user response
func (m *LoginUserResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this login user response based on context it is used
func (m *LoginUserResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *LoginUserResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *LoginUserResponse) UnmarshalBinary(b []byte) error {
	var res LoginUserResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}